+++ config/page-builder.php
@@ -24,14 +24,14 @@ return [
         'blockable'         => 'pagebuilder_blockables',
     ],
 
-    // The models are used in the package, you can use your own models, as long as they inherit the models below
+    // The models are used in the package; you can use your own models, as long as they inherit the models below
     'models'         => [
         'block'             => \Thinhnx\LaravelPageBuilder\Models\Block::class,
         'block_translation' => \Thinhnx\LaravelPageBuilder\Models\BlockTranslation::class,
         'blockable'         => \Thinhnx\LaravelPageBuilder\Models\Blockable::class,
     ],
 
-    // The transformers are used in the package, you can use your own transformers, as long as they inherit the transformers below
+    // The transformers are used in the package; you can use your own transformers, as long as they inherit the transformers below
     'transformers'   => [
         'block'             => \Thinhnx\LaravelPageBuilder\Transformers\BlockTransformer::class,
         'block_translation' => \Thinhnx\LaravelPageBuilder\Transformers\BlockTranslationTransformer::class,