+++ resources/assets/js/page-builder.js
@@ -1,102 +1,131 @@
 // Helper function
 function b64DecodeUnicode(str) {
-  return decodeURIComponent(atob(str).split('').map(function (c) {
-    return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
-  }).join(''));
+  return decodeURIComponent(atob(str)
+    .split('')
+    .map(function(c) {
+      return '%' + ('00' + c.charCodeAt(0)
+        .toString(16))
+        .slice(-2);
+    })
+    .join(''));
 }
 
-// Initialize Sortable for available block list
-function initBlockList() {
-  const blockList = document.getElementById("block-list");
-  Sortable.create(blockList, {
-    group: {name: "blocks", pull: "clone", put: false},
-    sort: false,
-  });
+// Initialize Sortable for the available blocklist
+function initBlockList(pageBuilderId) {
+  const blockList = document.getElementById(`block-list-${pageBuilderId}`);
+
+  if (blockList) {
+    Sortable.create(blockList, {
+      group: {
+        name: "blocks",
+        pull: "clone",
+        put: false
+      },
+      sort: false,
+    });
+  }
 }
 
 // Initialize Sortable for each language's editor
-async function initEditors() {
-  document.querySelectorAll(".sortable-container").forEach((container) => {
-    Sortable.create(container, {
-      group: "blocks",
-      animation: 150,
-      onAdd: addBlockToColumn,
-      onEnd: () => {
-        updateBlocksInput(container.getAttribute("data-locale"));
-      },
+async function initEditors(pageBuilderId) {
+  const pageBuilder = document.getElementById(pageBuilderId);
+
+  pageBuilder.querySelectorAll(".sortable-container")
+    .forEach((container) => {
+      Sortable.create(container, {
+        group: "blocks",
+        animation: 150,
+        onAdd: addBlockToColumn,
+        onEnd: () => {
+          updateBlocksInput(pageBuilderId, container.getAttribute("data-locale"));
+        },
+      });
     });
-  });
 }
 
-// Add block to column of layout
+// Add a block to the column of layout
 async function addBlockToColumn(evt) {
   let block = evt.item;
+  const pageBuilderId = block.getAttribute("data-page-builder-id");
   const type = block.getAttribute("data-type");
   const is_layout = block.getAttribute("data-is-layout");
-  block.innerHTML = await createBlockEditor(type);
+  block.innerHTML = await getBlockHtml(type);
   block.classList.remove("block");
 
   if (is_layout) {
-    block.querySelectorAll(".sortable-column").forEach((column) => {
-      Sortable.create(column, {
-        group: "blocks",
-        onAdd: addBlockToColumn,
-        onEnd: () =>
-          updateBlocksInput(container.getAttribute("data-locale")),
+    block.querySelectorAll(".sortable-column")
+      .forEach((column) => {
+        Sortable.create(column, {
+          group: "blocks",
+          onAdd: addBlockToColumn,
+          onEnd: () =>
+            updateBlocksInput(
+              pageBuilderId,
+              evt.to.closest(".sortable-container")
+                .getAttribute("data-locale")
+            ),
+        });
       });
-    });
   }
 
-  block.querySelector(".remove-block").addEventListener("click", function () {
-    block.remove();
-    updateBlocksInput(
-      evt.to.closest(".sortable-container").getAttribute("data-locale")
-    );
-  });
+  block.querySelector(".remove-block")
+    .addEventListener("click", function() {
+      block.remove();
+      updateBlocksInput(
+        pageBuilderId,
+        evt.to.closest(".sortable-container")
+          .getAttribute("data-locale")
+      );
+    });
 
   block
     .querySelector(".block-content")
     ?.addEventListener("input", () =>
       updateBlocksInput(
-        evt.to
-          .closest(".sortable-container")
+        pageBuilderId,
+        evt.to.closest(".sortable-container")
           .getAttribute("data-locale")
       )
     );
 }
 
-// Create HTML for block editor
-async function createBlockEditor(type, content = "") {
+// Get HTML of the block from the server
+async function getBlockHtml(type, content = "") {
   let html = null;
-  await fetch("/page-builder/render-block", {
+  await fetch(route("page-builder.render-block"), {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
-      "X-CSRF-TOKEN": document.querySelector('meta[name="csrf-token"]').content,
+      "X-CSRF-TOKEN": $('meta[name="csrf-token"]')
+        .attr("content"),
       "X-Requested-With": "XMLHttpRequest",
     },
-    body: JSON.stringify({type: type, content: content}),
+    body: JSON.stringify({
+      type: type,
+      content: content
+    }),
   })
     .then((response) => response.text())
     .then((text) => {
-      html = b64DecodeUnicode(JSON.parse(text).data);
+      html = b64DecodeUnicode(JSON.parse(text)
+        .data);
     });
 
   return html;
 }
 
-// Update blocks data into hidden input
-function updateBlocksInput(locale) {
-  const container = document.querySelector(
-    `.sortable-container[data-locale="${locale}"]`
-  );
+// Update blocks data into the hidden input
+function updateBlocksInput(pageBuilderId, locale) {
+  const pageBuilder = document.getElementById(pageBuilderId);
+  const container = pageBuilder.querySelector(`.sortable-container[data-locale="${locale}"]`);
   let blocks = getBlocksDataFromContainer(container);
 
-  document.getElementById(`blocks-${locale}`).value = JSON.stringify(blocks);
+  document.getElementById(`blocks-${locale}-${pageBuilderId}`)
+    .value = JSON.stringify(blocks);
   return blocks;
 }
 
-// Get blocks data from container
+// Get blocks data from a container
 function getBlocksDataFromContainer(container, column_index = 0) {
   let blockEditors = [];
   let blocks = [];
@@ -112,12 +141,15 @@ function getBlocksDataFromContainer(container, column_index = 0) {
 
   for (const block of blockEditors) {
     let children = [];
-    const type = block.closest("[data-type]").getAttribute("data-type");
+    const type = block.closest("[data-type]")
+      .getAttribute("data-type");
     const is_layout = block
       .closest("[data-type]")
       .getAttribute("data-is-layout");
-    const id = block.closest("[data-type]").getAttribute("data-id");
-    let content = block.querySelector(".block-content")?.value || "";
+    const id = block.closest("[data-type]")
+      .getAttribute("data-id");
+    let content = block.querySelector(".block-content")
+      ?.value || "";
 
     if (is_layout) {
       let row = block.querySelector(".row");
@@ -149,43 +181,48 @@ function getBlocksDataFromContainer(container, column_index = 0) {
 }
 
 // Initialize blocks from database data
-async function initBlocksFromData(locale, blocksData) {
-  const container = document.querySelector(
+async function initBlocksFromData(pageBuilderId, locale, blocksData) {
+  const pageBuilder = document.getElementById(pageBuilderId);
+  const container = pageBuilder.querySelector(
     `.sortable-container[data-locale="${locale}"]`
   );
-  await createBlockFromData(locale, blocksData, container);
+  await createBlockFromData(pageBuilderId, locale, blocksData, container);
 }
 
 async function createBlockFromData(
+  pageBuilderId,
   locale,
   blocksData,
-  container,
-  colIndex = 0
+  container
 ) {
   for (let block of blocksData) {
     const div = document.createElement("div");
+    div.setAttribute("data-page-builder-id", pageBuilderId);
     div.setAttribute("data-type", block.block.type);
     div.setAttribute("data-id", block.block_id);
     div.setAttribute("data-is-layout", block.block.is_layout);
-    div.innerHTML = await createBlockEditor(
+    div.innerHTML = await getBlockHtml(
       block.block.type,
       block.content
     );
     container.appendChild(div);
 
-    div.querySelector(".remove-block").addEventListener("click", function () {
-      div.remove();
-      updateBlocksInput(locale);
-    });
-    div.querySelector(".block-content")?.addEventListener("input", () =>
-      updateBlocksInput(locale)
-    );
+    div.querySelector(".remove-block")
+      .addEventListener("click", function() {
+        div.remove();
+        updateBlocksInput(pageBuilderId, locale);
+      });
+    div.querySelector(".block-content")
+      ?.addEventListener("input", () =>
+        updateBlocksInput(pageBuilderId, locale)
+      );
 
     if (block.block.is_layout) {
       let colIndex = 0;
       for (let column of div.querySelectorAll(".sortable-column")) {
         if (block.children.length) {
           await createBlockFromData(
+            pageBuilderId,
             locale,
             block.children.filter(
               (item) => item.column_index === colIndex
@@ -196,77 +233,89 @@ async function createBlockFromData(
         colIndex++;
       }
 
-      div.querySelectorAll(".sortable-column").forEach(
-        (column, index) => {
-          Sortable.create(column, {
-            group: "blocks",
-            onAdd: addBlockToColumn,
-            onEnd: () =>
-              updateBlocksInput(
-                container.getAttribute("data-locale")
-              ),
-          });
-        }
-      );
+      div.querySelectorAll(".sortable-column")
+        .forEach(
+          (column) => {
+            Sortable.create(column, {
+              group: "blocks",
+              onAdd: addBlockToColumn,
+              onEnd: () =>
+                updateBlocksInput(
+                  pageBuilderId,
+                  locale
+                ),
+            });
+          }
+        );
     }
   }
 }
 
 // Process preview
-function initPreview(callback = () => {
-}) {
+function initPreview(pageBuilderId, callback = () => {}) {
+  const pageBuilder = document.getElementById(pageBuilderId);
   let currentLocale = null;
-  document.querySelectorAll(".preview-btn").forEach((btn) => {
-    btn.addEventListener("click", function () {
-      currentLocale = this.getAttribute("data-locale");
-      const blocks = updateBlocksInput(currentLocale);
-      const iframe = document.getElementById("preview-iframe");
 
-      fetch("/page-builder/preview", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          "X-CSRF-TOKEN": document.querySelector('meta[name="csrf-token"]').content,
-          "X-Requested-With": "XMLHttpRequest",
-        },
-        body: JSON.stringify({
-          locale: currentLocale,
-          blocks: blocks,
-          context: callback(),
-        }),
-      })
-        .then((response) => response.text())
-        .then((text) => {
-          iframe.srcdoc = b64DecodeUnicode(JSON.parse(text).data);
-          updateIframeSize("desktop");
+  pageBuilder.querySelectorAll(".preview-btn")
+    .forEach((btn) => {
+      btn.addEventListener("click", function() {
+        currentLocale = this.getAttribute("data-locale");
+        const blocks = updateBlocksInput(pageBuilderId, currentLocale);
+        const iframe = document.getElementById("preview-iframe");
+
+        fetch(route("page-builder.preview"), {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+            "X-CSRF-TOKEN": document.querySelector(
+              'meta[name="csrf-token"]'
+            )
+              .content,
+            "X-Requested-With": "XMLHttpRequest",
+          },
+          body: JSON.stringify({
+            locale: currentLocale,
+            blocks: blocks,
+            context: callback(),
+          }),
         })
-        .catch((error) => console.error("Error:", error));
+          .then((response) => response.text())
+          .then((text) => {
+            iframe.srcdoc = b64DecodeUnicode(JSON.parse(text)
+              .data);
+            updateIframeSize("desktop");
+          })
+          .catch((error) => console.error("Error:", error));
+      });
     });
-  });
 
   document
     .querySelectorAll('input[name="device-select"]')
-    .forEach(function (device) {
-      device.addEventListener("change", function () {
+    .forEach(function(device) {
+      device.addEventListener("change", function() {
         updateIframeSize(
           document.querySelector(
             'input[name="device-select"]:checked'
-          ).value
+          )
+            .value
         );
       });
     });
 }
 
 // Process form submission
-function initFormSubmit(locales, formId) {
-  document.getElementById(formId).addEventListener("submit", function () {
-    locales.forEach((locale) => updateBlocksInput(locale));
-    document
-      .querySelectorAll('input[name="device-select"]')
-      .forEach(function (device) {
-        device.disabled = true;
-      });
-  });
+function initFormSubmit(pageBuilderId, locales) {
+  const pageBuilder = document.getElementById(pageBuilderId);
+
+  pageBuilder.closest('form')
+    .addEventListener("submit", function() {
+      locales.forEach((locale) => updateBlocksInput(pageBuilderId, locale));
+      document
+        .querySelectorAll('input[name="device-select"]')
+        .forEach(function(device) {
+          device.disabled = true;
+        });
+    });
 }
 
 // Adjust the size of the iframe
@@ -291,21 +340,21 @@ function updateIframeSize(device) {
   }
 }
 
-// Initialize the entire editor
-async function initBlockEditor(
+// Initialize the Page Builder
+async function initPageBuilder(
+  pageBuilderId,
   initialBlocks = {},
-  getContextCallback = () => {
-  }
+  getContextCallback = () => {}
 ) {
-  initBlockList();
-  await initEditors();
-  initPreview(getContextCallback);
-  initFormSubmit(Object.keys(initialBlocks), window.formId);
+  initBlockList(pageBuilderId);
+  await initEditors(pageBuilderId);
+  initPreview(pageBuilderId, getContextCallback);
+  initFormSubmit(pageBuilderId, Object.keys(initialBlocks));
 
   // Initialize blocks from database data
   for (const locale in initialBlocks) {
     if (initialBlocks[locale].length > 0) {
-      await initBlocksFromData(locale, initialBlocks[locale]);
+      await initBlocksFromData(pageBuilderId, locale, initialBlocks[locale]);
     }
   }
 }